# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Community Donation Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List all active donations
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a donation listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get a donation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    patch:
      operationId: updateDonation
      tags: [donations]
      summary: Update a donation listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationUpdate"
      responses:
        "200":
          description: Updated donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    delete:
      operationId: deleteDonation
      tags: [donations]
      summary: Delete a donation listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /donations/{id}/complete:
    patch:
      operationId: completeDonation
      tags: [donations]
      summary: Mark donation as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"

  # ── RECEIVER REQUESTS ─────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/verify:
    patch:
      operationId: verifyUser
      tags: [admin]
      summary: Verify a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend or unsuspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── USER SCHEMAS ────────────────────────────────────────────────────
    User:
      type: object
      required: [id, fullName, email, phone, role, isVerified, trustScore, trustLevel, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [donor, receiver, admin]
        isVerified:
          type: boolean
        trustScore:
          type: integer
        trustLevel:
          type: string
          enum: [new, trusted, highlyTrusted, champion]
        totalDonations:
          type: integer
        averageRating:
          type: ["number", "null"]
        profilePhoto:
          type: ["string", "null"]
        village:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
...
    UserRegistration:
      type: object
      required: [fullName, email, phone, password, role]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [donor, receiver]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    Donation:
      type: object
      required: [id, donorId, donorName, category, itemName, quantity, condition, status, radiusKm, createdAt]
      properties:
        id:
          type: integer
        donorId:
          type: integer
        donorName:
          type: string
        donorVerified:
          type: boolean
        donorTrustScore:
          type: integer
        donorTrustLevel:
          type: string
        category:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
        brand:
          type: ["string", "null"]
        purchaseYear:
          type: ["integer", "null"]
        condition:
          type: string
          enum: [likeNew, excellent, good, fair, poor]
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        village:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        radiusKm:
          type: integer
        status:
          type: string
          enum: [active, matched, completed, cancelled]
        distance:
          type: ["number", "null"]
        createdAt:
          type: string
    DonationInput:
      type: object
      required: [category, itemName, quantity, condition, radiusKm]
      properties:
        category:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
        brand:
          type: string
        purchaseYear:
          type: integer
        condition:
          type: string
          enum: [likeNew, excellent, good, fair, poor]
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radiusKm:
          type: integer
    DonationUpdate:
      type: object
      properties:
        itemName:
          type: string
        quantity:
          type: integer
        condition:
          type: string
        radiusKm:
          type: integer
        status:
          type: string

    # ── REQUEST SCHEMAS ───────────────────────────────────────────────────
...
    SuspendInput:
      type: object
      required: [suspended]
      properties:
        suspended:
          type: boolean
        reason:
          type: string