# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Community Donation Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';
import type { UserTrustLevel } from './userTrustLevel';

export interface User {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  role: UserRole;
  isVerified: boolean;
  trustScore: number;
  trustLevel: UserTrustLevel;
  totalDonations?: number;
  /** @nullable */
  averageRating?: number | null;
  /** @nullable */
  profilePhoto?: string | null;
  /** @nullable */
  village?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  state?: string | null;
  status: UserStatus;
  createdAt: string;
}